updating total sales to department

+++ bamazonSchema.sql
@@ -15,5 +15,5 @@ CREATE TABLE departments (
 department_id INT PRIMARY KEY AUTO_INCREMENT NOT NULL,
 department_name VARCHAR(50) NOT NULL,
 over_head_costs DECIMAL(10,2),
-total_sales INT
+total_sales DECIMAL(10,2)
 );
